updates for JS_pp200_signal

+++ submit/JS_pp200_signal/cemc_geo_fix/cemc_hits_pass2/condor/restart_held.sh
@@ -1,9 +1,20 @@
 #! /usr/bin/bash
+if [ $# -eq 0 ]
+  then
+    echo "No arguments supplied"
+    exit 1
+fi
+echo $1
+variable=$1
 [ -f bla ] && rm bla
-condor_q | grep ' H ' | grep run_cemc_pass2_geo_fix_0_20fm.sh > bla
+condor_q | grep ' H ' | grep run_cemc_pass2_geo_fix_js.sh  | grep ${variable} > bla
 #condor_q | grep ' H ' | grep run_pileup.sh | grep sHijing_0_20 > bla
 [ ! -s bla ] && exit 0
 for i in `cat bla| awk '{print $1}'`; do condor_rm $i; done
-#for i in `cat bla | awk '{print $12}' | awk -F- '{print $3}' | awk -F. '{print "log/condor-0000000006-"$1".job"}'`; do sed -i 's/run_pileup.sh/run_pileup_0_20fm.sh/' $i; echo $i; condor_submit $i; done
-for i in `cat bla | awk '{print $12}' | awk -F- '{print $3}' | awk -F. '{print "log/condor-0000000006-"$1".job"}'`; do echo $i; condor_submit $i; done
-[ -f bla ] && rm bla
+[ -f tmplist ] && rm tmplist
+for i in `cat bla | awk '{print $12}' | awk -F- '{print $3}' | awk -F. '{print "-0000000006-"$1".job"}'`; do echo $i >> tmplist ; done
+[ -f sedlist ] && rm sedlist
+for i in `cat tmplist`; do echo log/${variable}/condor_${variable}$i >> sedlist; done
+#for i in `cat sedlist`; do  sed -i 's/12288MB/24576MB/' $i; echo $i; done
+for i in `cat sedlist`; do condor_submit $i; done
+
